requirements.txt
alembic==1.10.2
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
comm==0.1.2
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyzmq==25.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
stack-data==0.6.2
starlette==0.26.0.post1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn==0.21.0
wcwidth==0.2.6
zipp==3.15.0
